Refactor error modules
