+++ http-wrapper.gemspec
@@ -38,10 +38,10 @@ Gem::Specification.new do |spec|
   spec.add_dependency "oj", "~> 3.16", ">= 3.16.1"
   spec.add_dependency "rspec", "~> 3.11", ">= 3.11.0"
 
-  spec.add_development_dependency 'bundler'
-  spec.add_development_dependency 'byebug', '~> 11.1', '>= 11.1.3'
-  spec.add_development_dependency 'rake', '~> 13.1'
-  spec.add_development_dependency 'rspec-rails', '~> 6.0', '>= 6.0.3'
+  spec.add_development_dependency "bundler"
+  spec.add_development_dependency "byebug", "~> 11.1", ">= 11.1.3"
+  spec.add_development_dependency "rake", "~> 13.1"
+  spec.add_development_dependency "rspec-rails", "~> 6.0", ">= 6.0.3"
 
   # For more information and examples about making a new gem, check out our
   # guide at: https://bundler.io/guides/creating_gem.html
