+++ lib/http/wrapper.rb
@@ -6,7 +6,7 @@ require "http/wrapper/version"
 require "http/wrapper/http_status_codes"
 require "http/wrapper/api_exceptions"
 require "http/wrapper/configuration"
-require "http/wrapper/error_mapping"
+require "http/wrapper/error_handling"
 require "http/wrapper/request_handler"
 require "http/wrapper/response_handler"
 require "http/wrapper/client"
