+++ lib/http/wrapper/response_handler.rb
@@ -1,6 +1,10 @@
+# frozen_string_literal: true
+
 module Http
   module Wrapper
-      module ResponseHandler
+    module ResponseHandler
+      private
+
       def handle_response
         parsed_response = Oj.load(@response.body)
 
@@ -9,22 +13,8 @@ module Http
         raise error_class, "Code: #{@response.status}, response: #{@response.body}"
       end
 
-      def error_class
-        ERROR_MAPPING[@response.status] || UnknownStatusError.new(@response.status)
-      end
-
       def response_successful?
         SUCCESSFUL_STATUS.include?(@response.status)
-      end
-
-      def api_requests_quota_reached?
-        @response.body.match?(API_REQUESTS_QUOTA_REACHED_MESSAGE)
-      end
-
-      def forbidden_error(response)
-        return ApiExceptions.const_get("ApiRequestsQuotaReachedError").new if api_requests_quota_reached?(response)
-
-        ApiExceptions.const_get("ForbiddenError").new
       end
     end
   end
